+++ packages/renderer-ui/src/features/toolbars/index.ts
@@ -1,3 +1,5 @@
+export { default as ToolbarButton } from './ui/ToolbarButton.vue';
+export { default as ToolbarColorButton } from './ui/ToolbarColorButton.vue';
 export { default as FloatingToolbar } from './ui/FloatingToolbar.vue';
 export { default as ToolbarGroup } from './ui/ToolbarGroup.vue';
-export type { FloatingToolbarPosition, FloatingToolbarProps } from './definitions/toolbar';
+export type { FloatingToolbarPosition, FloatingToolbarProps, ToolbarButtonProps, ToolbarColorButtonProps } from './definitions/toolbar';